Move supervisor into ibc-relayer crate
