+++ relayer/src/lib.rs
@@ -26,5 +26,6 @@ pub mod light_client;
 pub mod link;
 pub mod macros;
 pub mod relay;
+pub mod supervisor;
 pub mod transfer;
 pub mod util;
